+++ src/main/java/pl/betoncraft/betonquest/JoinQuitListener.java
@@ -57,7 +57,7 @@ public class JoinQuitListener implements Listener {
         // display changelog message to the admins
         if (event.getPlayer().hasPermission("betonquest.admin")
                 && new File(BetonQuest.getInstance().getDataFolder(), "CHANGELOG.md").exists()) {
-            Config.sendNotify(PlayerConverter.getID(event.getPlayer()), "changelog", null, "changelog,info");
+            Config.sendNotify(null, PlayerConverter.getID(event.getPlayer()), "changelog", null, "changelog,info");
         }
         if (Journal.hasJournal(playerID)) {
             playerData.getJournal().update();