+++ src/main/java/pl/betoncraft/betonquest/notify/ChatNotifyIO.java
@@ -4,6 +4,7 @@ import org.bukkit.entity.Player;
 import pl.betoncraft.betonquest.utils.Utils;
 
 import java.util.Collection;
+import java.util.HashMap;
 import java.util.Map;
 
 /**
@@ -16,11 +17,10 @@ public class ChatNotifyIO extends NotifyIO {
     }
 
     @Override
-    public void sendNotify(final String message, final Collection<? extends Player> players) {
-        for (final Player player : players) {
-            player.sendMessage(Utils.format(message));
+    public void sendNotify(final HashMap<Player, String> playerMessages) {
+        for (final Map.Entry<Player, String> entry : playerMessages.entrySet()) {
+            entry.getKey().sendMessage(Utils.format(entry.getValue()));
         }
-
-        sendNotificationSound(players);
+        sendNotificationSound(playerMessages.keySet());
     }
 }