+++ src/main/java/pl/betoncraft/betonquest/notify/NotifyIO.java
@@ -3,18 +3,21 @@ package pl.betoncraft.betonquest.notify;
 import org.bukkit.Bukkit;
 import org.bukkit.Sound;
 import org.bukkit.entity.Player;
+import pl.betoncraft.betonquest.BetonQuest;
+import pl.betoncraft.betonquest.config.ConfigPackage;
 import pl.betoncraft.betonquest.utils.LogUtils;
+import pl.betoncraft.betonquest.utils.PlayerConverter;
 
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Set;
 import java.util.logging.Level;
 
 /**
  * Used to display messages to a player
  * <p>
- * Data Valuues:
+ * Data Values:
  * * sound: {sound_name} - What sound to play
  */
 public abstract class NotifyIO {
@@ -45,8 +48,9 @@ public abstract class NotifyIO {
         return this;
     }
 
-    public void sendToAll(final String message) {
-        sendNotify(message, Bukkit.getServer().getOnlinePlayers());
+    public void sendToAll(final String packName, final String message) {
+        final Collection<? extends Player> players = Bukkit.getServer().getOnlinePlayers();
+        sendNotify(packName, message, players.toArray(new Player[players.size()]));
     }
 
     /**
@@ -55,16 +59,29 @@ public abstract class NotifyIO {
      * @param players Players to show
      */
 
-    public void sendNotify(final String message, final Player... players) {
-        sendNotify(message, Arrays.asList(players));
+    public void sendNotify(final String packName, final String message, final Player... players) {
+        final HashMap<Player, String> playerMessages = new HashMap<>();
+        if (packName == null) {
+            for (final Player player : players) {
+                playerMessages.put(player, message);
+            }
+        } else {
+            for (final String variable : BetonQuest.resolveVariables(message)) {
+                for (final Player player : players) {
+                    final String resolvedMessage = message.replace(variable, BetonQuest.getInstance().getVariableValue(packName, variable, PlayerConverter.getID(player)));
+                    playerMessages.put(player, resolvedMessage);
+                }
+            }
+        }
+        sendNotify(playerMessages);
     }
 
     /**
      * Show a notify to a collection of players
      */
-    public abstract void sendNotify(final String message, final Collection<? extends Player> players);
+    public abstract void sendNotify(final HashMap<Player, String> playerMessages);
 
-    protected void sendNotificationSound(final Collection<? extends Player> players) {
+    protected void sendNotificationSound(final Set<Player> players) {
         if (getData().containsKey("sound")) {
             for (final Player player : players) {
                 try {