+++ src/main/java/pl/betoncraft/betonquest/notify/TitleNotifyIO.java
@@ -3,7 +3,7 @@ package pl.betoncraft.betonquest.notify;
 import org.bukkit.entity.Player;
 import pl.betoncraft.betonquest.utils.Utils;
 
-import java.util.Collection;
+import java.util.HashMap;
 import java.util.Map;
 
 /**
@@ -29,18 +29,17 @@ public class TitleNotifyIO extends NotifyIO {
     public TitleNotifyIO(final Map<String, String> data) {
         super(data);
 
-        fadeIn = Integer.valueOf(data.getOrDefault("fadein", "10"));
-        stay = Integer.valueOf(data.getOrDefault("stay", "70"));
-        fadeOut = Integer.valueOf(data.getOrDefault("fadeout", "20"));
+        fadeIn = Integer.parseInt(data.getOrDefault("fadein", "10"));
+        stay = Integer.parseInt(data.getOrDefault("stay", "70"));
+        fadeOut = Integer.parseInt(data.getOrDefault("fadeout", "20"));
         subTitle = getData().getOrDefault("subtitle", "").replace("_", " ");
     }
 
     @Override
-    public void sendNotify(final String message, final Collection<? extends Player> players) {
-        for (final Player player : players) {
-            player.sendTitle(Utils.format(message), Utils.format(subTitle), fadeIn, stay, fadeOut);
+    public void sendNotify(final HashMap<Player, String> playerMessages) {
+        for (final Map.Entry<Player, String> entry : playerMessages.entrySet()) {
+            entry.getKey().sendTitle(Utils.format(entry.getValue()), Utils.format(subTitle), fadeIn, stay, fadeOut);
         }
-
-        sendNotificationSound(players);
+        sendNotificationSound(playerMessages.keySet());
     }
 }